+++ database/migrations/2015_07_11_110339_create_planet_starting_resources_table.php
@@ -14,7 +14,7 @@ class CreatePlanetStartingResourcesTable extends Migration
     {
         Schema::create('planet_starting_resources', function (Blueprint $table) {
             $table->integer('resource_id')->unsigned();
-            $table->integer('stored')->unsigned();
+            $table->bigInteger('stored')->unsigned();
             $table->integer('abundance');
 
             $table->primary('resource_id');
